The first version of the project has been uploaded.
The first version of the project has been uploaded. This version has its own system for searching fraudulent resources on the built-in database, and also outputs the results of the check to the console. For convenience, various items of the test results were highlighted in different colors, pleasing to the eye.